Add files via upload
This is the first version of the program. It is working but it can definitely receive some improvements. It was a good project to learn how to manipulate JFrames in java, specifically manipulating their location and transparency. Also it was a good way to learn how to use taskbar icons, and manipulating files/directories in java. My vision of this app was to make it a widget with dynamically adjusting window size that would "open up" when focused on by the mouse cursor and give you an icon preview of all the applications in the specific folder, where you could launch them with just a click. Also some possible improvements include giving the user options to "open the directory" where the applications/ application shortcuts are saved, and giving the user the option to "add shortcuts" to the directory with some extra small buttons and a few clicks. For now i am focusing on other bigger projects though.